# server/contracts/openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "VoxProbe Sandbox E-Commerce Microservices API",
    "version": "1.0.0",
    "description": "Mock diagnostic sandbox API for AssemblyAI Voice Agent Hackathon contract verification."
  },
  "servers": [
    {
      "url": "http://localhost:3001",
      "description": "VoxProbe Local Sandbox API Server"
    }
  ],
  "paths": {
    "/api/orders/{id}": {
      "get": {
        "summary": "Retrieve order details by ID",
        "operationId": "getOrderById",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Order identifier (e.g. ORD-1042)"
          }
        ],
        "responses": {
          "200": {
            "description": "Order response schema",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": ["id", "status", "total", "currency"],
                  "properties": {
                    "id": {
                      "type": "string",
                      "pattern": "^ORD-\\d{4}$"
                    },
                    "status": {
                      "type": "string",
                      "enum": ["pending", "processing", "shipped", "delivered"]
                    },
                    "total": {
                      "type": "number",
                      "minimum": 0
                    },
                    "currency": {
                      "type": "string",
                      "enum": ["INR", "USD", "EUR"]
                    }
                  },
                  "additionalProperties": false
                }
              }
            }
          },
          "404": {
            "description": "Order not found"
          }
        }
      }
    },
    "/api/users/{id}": {
      "get": {
        "summary": "Retrieve user profile by ID",
        "operationId": "getUserById",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "User profile schema",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": ["id", "name", "email", "role"],
                  "properties": {
                    "id": {
                      "type": "string"
                    },
                    "name": {
                      "type": "string"
                    },
                    "email": {
                      "type": "string",
                      "format": "email"
                    },
                    "role": {
                      "type": "string",
                      "enum": ["admin", "developer", "customer"]
                    }
                  },
                  "additionalProperties": false
                }
              }
            }
          }
        }
      }
    },
    "/api/products/{id}": {
      "get": {
        "summary": "Retrieve product details by ID",
        "operationId": "getProductById",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Product schema",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": ["id", "title", "price", "in_stock"],
                  "properties": {
                    "id": {
                      "type": "string"
                    },
                    "title": {
                      "type": "string"
                    },
                    "price": {
                      "type": "number"
                    },
                    "in_stock": {
                      "type": "boolean"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/login": {
      "post": {
        "summary": "User authentication endpoint",
        "operationId": "loginUser",
        "responses": {
          "200": {
            "description": "Authentication token response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": ["token", "token_type", "expires_in"],
                  "properties": {
                    "token": {
                      "type": "string"
                    },
                    "token_type": {
                      "type": "string"
                    },
                    "expires_in": {
                      "type": "number"
                    }
                  }
                }
              }
            }
          }
        }
      }
    }
  }
}